feat: Add MicrosoftBotAPI for handling Microsoft Teams bot interactions and remove deprecated MicrosoftTeamsBotAPI
